ci(build): create the docker image and push it to `${CI_REGISTRY}`
We use the EOLE/ci-tools> template to build the docker image and push
it to the registry as `laboite-blog_front:git-${CI_COMMIT_SHORT_SHA}`.

We build on every branches except for `$STABLE_BRANCH` used only to
create the release tag and for `$TESTING_BRANCH` used only to create
the prerelease tag.